Comment
Hello people of Kiwi, first of all I would like to thank you to give me the chance to make this test.
I made the script as much better as I could, but I still think that with more time I would be able to improve it somehow. However,
this would make this task very long :)
What I did was find an online database with the money rates. I prepared the online data for be used and, finally I made two functions that  make the money conversion and the money symbol conversion.
As junior python programmer I am aware that I have plenty of things to learn. But with a proper training I would be able to do great things!!
Kind regards